enable partition based SelectType
Hi everybody!

Here attached is a patch that enables partition based SelectType (currently CR_Socket/CR_Core) in select/cons_res.

The patch requires that CR_ALLOCATE_FULL_SOCKET is enabled to work and also this patch from master branch: https://github.com/SchedMD/slurm/commit/cdf679d0158a246e7389a15b62f127e5142003fe

It should however be easy to change it to use the old #define if you want to.

We are currently testing this in our development system but will go into production later this spring based on needs from some of our users.

One thing that I noticed during the development of this is that if a new option is added to the slurm.conf that is not supported with an earlier version of slurm programs/libs that are compiled with the earlier version stops working due to complaining of errors in slurm.conf.

We have the CR_ALLOCATE_FULL_SOCKET patch in our production system and some programs linked with openmpi stop working for some of our users.

It might we wise to try require less reading of the slurm.conf from the core parts of slurm and try to put more reading/parsing of the config file from the plugins (and other modular parts of slurm).

Best regards,
Magnus