Permit backfill scheduler to continue executing after reliquishing locks
Jobs are not backfilled due to the fact that backfill does not finish the complete backlog of jobs in the queue before it's interrupted and starts all over again. We sometimes have lots of jobs in the queue of various sizes and users and even with idle nodes short job will not start because of this.  I have made a patch for backfill with a configuration option (bf_continue) to let backfill continue.