Make testing of node names more strict
If node_name2bitmap() is called with best_effort=false, then do
not attempt to match names with NodeHostName.

Without this change, a partition that contains a NodeHostName rather
that NodeName would be configured with the first one found. On a
front-end system, this would result in the partition's node_bitmap
being out of sync with the actual node positions.

To reproduce the problem, configure with --enable-multiple-slurmd
Then in slurm.conf, define something like this:
NodeName=foo[1-8] NodeHostName=bar ...
PartitionName=debug Nodes=bar,foo[1-8] ...