Add some Intel Phi documentation to SLURM FAQ, just a placeholder for now
Thanks for the feedback! I was actually planning contributing a more complete set of documentation for the SLURM web pages but have been super busy.
Feel free to copy the current documentation. I'll try to refine it a bit next week.
It might be better to leave the link out as we might move the documentation to the "proper" CSC website at some point.