Fix calculation of tasks per code
A simple one-line fix to the "_adjust_cpus_nppcu" function that I had
added.  I had added this function as part of the NPPCU functionality;
however it wasn't a problem until that squeue patch.  That was because
then squeue had been updated to use this function and in this one case
the default value for the internal variable "ntasks_per_core" wound up
not being the 0xffff (65535) that I previously had coded for (as in
"select/cons_res") but instead was 0.  Therefore, in that adjustment
function of mine, I simply added a second clause to the if-statment
where I check for the sentinel value that also checks whether it is 0.
This resolved the problem.  Because we do not usually use
"select/serial", I did not notice this.