PMI2 strcpy bug
This is the correction of the slurm client which otherwise fails the
PMI2_Init() step. This hidden bug was introduced in the standalone
client made from BULL and David's modification made it occur : a wrong
size (by 1 too short) was passed to strncpy for keys, then David
replaced strncpy by MPICH's MPIU_Strncpy forcing a terminal '\0'.

Bug 359