Move location when job container plugin is called from
Before a task is attached to a PAGG and the PAGG attached to a
reservation, no tracking is done.  Any files and directories the task
creates at such a time will not be tracked.

When a task is removed from a PAGG (via job_detachpid()), its current
cleanup state is dropped, and no further files it creates will be tracked.

Put another way, job_detachpid() will cancel any cleanup state accrued by
that task, and make it as if that task had never been subject to CNCU.

Once the creating task exits while subject to CNCU, the CNCU-tracking of
their files is inherited by the PAGG, and then by the reservation when the
PAGG exits.

So if the slurmstepd task had spawned child tasks, and those tasks create
files and exit while still attached to the PAGG attached to the
reservation, then those files will be cleaned up when the reservation
ends.

Paul