Fix default share behavior for select/cons_res
Fix the sharing options for select/cons_res to adjust to documentation.

Specifically, given a partition is set to allow over-subscription of
resources and a user does not request exclusive access to resources,
set the job sharing properties as NODE_CR_AVAILABLE as the default option
regarless of the presence of the --share option at submission time.