Reservation with CoreCnt: Avoid possible invalid memory reference.
If reservation create request included a CoreCnt value and more
nodes are required than configured, the logic in select/cons_res
could go off the end of the core_cnt array. This patch adds a
check for a zero value in the core_cnt array, which terminates
the user-specified array.