acct_gather_energy/ipmi - Add delay before retry on read error.
If slurmd fails to get IPMI value, then I propose to force to wait 1 second instead of asking BMC again. (Part 3/4 of the patch).
If IPMI init fails when slurmd forces to update the value, then we should not update the value. (Part 4/4 of the patch
Part 1/4 and 2/4 add a security in IPMI init because the function can be call several time.
This force to return SLURM_FAILURE if the first call failed, since the other call will not do anything.

bug 469