Corrections to advanced reservation logic with overlapping jobs.
This might have worked fine for core reservations or when there
are sufficient idle nodes to use, the the select_g_resv_test()
function clears the node bitmap for nodes that it can not use
and the reservation create logic did not restore that bitmap
after a failed resource selection attempt. This logic restores
the node bitmap on a failed call to select_g_resv_test() so we
can add nodes to the bitmap of available nodes rather than having
it repeatedly cleared.
The logic also adds some performance enhancements that I will
add to in the next commit.