Check node name on batch job completion
Add a check to make sure that the job completion RPC from a
slurmstepd match that node that the batch job is running on.
This would not be the case of for a job started on a node
if that node's slurmd fails, but the slurmstepd keeps running.
The job could then be requeued and generate a completion RPC
from both slurmstepd daemons (one per node). This logic will
ignore the job complete RPC from the node NOT currently
running the batch job.