Make sure node limits get assessed if no node count was given in request.
i.e. salloc -n32 doesn't request the number of nodes and with the previous
code if this request used 4 nodes and only 1 was left in GrpNodes it
would just run with no issue since we were checking things before we
selected how many nodes it ran on.

Now we check this afterwards so we always check the limits on how many
nodes, cpus and how much memory is to be used.