select/cray step cleanup
If the slurm.conf changes to remove nodes from a job step in the
process of running Node Health Check, the step record would be
purged. Then when the NHC completes, it's step record pointer
would be invalid.
The change here will prevent such a job step record from being
purged.
This was found while investigating bug 612, but would be a very
unusual condition and does not fix this bug.