Tweak test timing
In test15.17, an salloc spawns an sbatch within that same job
allocation then exits. This results in a race condition which
can cause the terminate job RPC from slurmctld to slurmd to
briefly hang and mark the node non-responsive, causing later
tests that may need the node to fail (timeout waiting for
job allocation). A sleep is added to the salloc here to give
the batch job a chance to begin and clean up the job allocation
quickly.