Fix sview abort when adding/removing columns
When adding or removing columns to most data types (jobs, partitions,
nodes, etc.) on some system types an abort is generated. This appears
to be because when columns displayed change, on some systems that
changes the address of "model", while on others the address does not
change (like my laptops). This fix explicitly sets the last_model to
NULL when the columns are changed rather than relying upon the data
structure's address to change.