Remove WAIT_SCHED_TIMEOUT
This actually causes more problems since there are a bunch of
places where we set the reason only if it was previously NO_REASON,
but we want to treat this the same way, i.e. untested. Adding
logic everywhere to clear and reset this value seem to create
too much overhead for very little value.