Clear completing node bitmap on job requeue
When a batch job requeue completes, clear it's bitmap of completing
nodes. If the bitmap were to persist, and nodes are added/remove in
slurm.conf, and "scontrol reconfigure" is executed, and one of the
bits in that bitmap now point to a DOWN node, avoid killing the job.
bug 805