Speed up shutdown process
If shutdown of the slurmctld daemon is in progress, then stop
trying to schedule jobs or process reconfigure requests. These
are the only operations that take a significant amount of time
and only service to slow down the shutdown process. We want the
daemon to stop processing incoming RPCs and save state as soon
as possible.