Allocate step part of job's allocation
If a job allocates whole nodes (with --core-spec or --exclusive
option) and launches a step within that allocation all in a single
command (the srun creates the allocation and step at the same time)
and requests a specific CPU count (e.g. "-ntasks=# --cpus-per-task=#")
then allocate the job step only the requested CPU count, which may
be less than the job's allocation.
Bug 843