fix for job submit hang
If the job to be suspended can not be found, then proceed with the
suspend RPC after 3 seconds, Previous logic would hang indefinitely.
The job would not be found if it was launched, then the slurmd restarted.
The record keeping for launched jobs after slurmd restarts is fixed
in a separate patch.