scontrol show job report correct CPU_IDs
For "scontrol --details show job" report the correct CPU_IDs when thre are
multiple threads per core (we are translating a core bitmap to CPU IDs).
This is an enhancement of commit 83d626ca45032d71f3847a9a92e9a21fac5e8031
so the node table is only loaded once for the entire job table.
bug 850