Move job array task ID string format
Move the formatting of a job array's task ID string from slurmctld
to the client to reduce system overhead. There is logic in place
to translate the full bitmap into a string, but in the worst case
this can take roughly one minute (e.g. 1-1000000:2,2) so by default
only the first 1024 bytes in the string are set. There is logic in
the code to translate the full string, but commented out.