CPU frequency set race condition fix
Fix race condition in CPU frequency set with job preemption.
When the preemptor job completed, it would notify the srun, which
would notify the slurmctld, which could resume a preempted job.
That preempted job could reset the CPU frequency before the
preemptor. This change has the slurmstepd resetting a job's
CPU frequency prior to notifying srun of completion, which
eliminates the race condition.
bug 1011