refactor job step delete logic
Call delete_step_records() before clearing the job's JOB_COMPLETING
state flag. This would make a difference in the case of jobs automatically
requeued based upon their exit code, but probably not in other cases.
Also in the select plugins, check not only for a job state of JOB_COMPLETING,
but also FINISHED states. In either case, we are not in a position to
gracefully clean up the step.