fix gang sched for jobs in multiple partition
Fix gang scheduling for jobs submitted to multiple partitions.
Previous logic assumed the job's "partition" field contained a
single partition name, that in which the job is running. That
was recently changed in order to support job's being requeued,
which we want to be runable in all of it's valid partitions.