Correct step's node allocation
Under some conditions a step may be allocated fewer nodes than requested. Specifically, if the job uses the --exclusive --nodelist and in some fashion a higher node count, it may only be allocated the nodes in the specified node list. For example:
$ srun  -n2 --exclusive --ntasks-per-node=1 --cpus-per-task=4 -w smd1 hostname
smd1
smd1

This problem appears to only exist in version 14.11 and only applies to the step allocation, not the job allocation.