Initially, the collector and backup ports are both that of slurmctld. As I iterate through the hierarchy, they get changed to that of the respective slurmd nodes.
I initially used slurm_get_slurmd_port() but it returns the value of
SlurmdPort from slurm.conf. This is the same value for all emulated
slurmd.

What I need is the port assigned to each the slurmd from the node
definition record. For example NodeName=trek[000-054] Port=[18000-18054],
I need 18000,... not 6818.

This is what this patch does, it gets the port from the config instead of
the default from the api.