Fix for job array tracking with select/linear
We found that the job array's job id is being added to the allocated
nodes of the running jobs instead of the individual job id's of the
jobs in the array.

We also found that putting job_array_post_sched() before
select_g_job_begin() in node_scheduler.c:node_select() and removing
the check for IS_JOB_RUNNING in job_array_post_sched() makes it work.
job_array_post_sched() is only called from node_select() so it
should be safe to remove the IS_JOB_RUNNING check.

see bug 1120