Fix for resuming powered down node
This permits a sys admin to power down a node that should already
be powered down, but avoids setting the NO_RESPOND bit in the
node state. Doing so under some conditions prevented the node from
being scheduled. The downside is that the node could possibly be
allocated when it really isn't ready for use.