Don't invoke spank prolog/epilog if no config file
Note that PlugStackConfig defaults to plugstack.conf in the same
directory as slurm.conf. The added logic tests if the file actually
exists (using stat) and if not found then do not fork/exec slurmstepd
to invoke the spank prolog/epilog. This saves about 14msec on startup
and 14msec on shutdown if no spank plugins are configured. It also
eliminates some possible failures (e.g. if fork() fails, or the
slurmstepd processes can not exec()). This logic also caches the
PlugStackConfig value and reads it again on reconfigure, but avoid
reading the value for each job.
bug 982