Improve config preservation on reconfig
The logic was flawed to preserve BurstBufferType on slurmctld reconfig
In addition for many plugins, one could go from no value to some
value (i.e. a change in plugin) without reconfiguring. In this case,
the slurmctld logic would not match what is in slurm.conf or reported
by "scontrol show config".