Remove CR_ALLOCATE_FULL_SOCKET option.  It is now the default.
Before without this option accounting would not be correct unless the
job was allocating enough resources to fill up the socket.  Making this
the default makes it so the entire socket is allocated to job similar to
the way CR_CORE works with allocating the entire core to the job even if
they don't allocate the whole core.