Prevent vestigial job array record
If a master job array record is complete, then consider all pending
tasks as also complete. This problem happens when a master job array
record is pending (has pending tasks) and is cancelled. The result
previously was a job record not visible to squeue/scontrol, but occupying
memory.
The same type of problem happened with respect to a dependency on a job
array which was cancelled.